refactor(tools): refactor workspace generators, bump project dependencies
- [x] de-duplicate workspace generators code;
- [x] update project dependencies without breaking changes;
