+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nx-ng-starter",
-  "version": "3.0.7",
+  "version": "3.1.0",
   "private": true,
   "description": "Monorepo starter: Nx, Angular, Angular Elements, Electron, NodeJS, NestJS, Firebase.",
   "license": "MIT",
@@ -195,7 +195,7 @@
     "graphql": "16.6.0",
     "graphql-subscriptions": "2.0.0",
     "graphql-tools": "8.3.4",
-    "graphql-upload": "16.0.1",
+    "graphql-upload": "16.0.2",
     "grpc-web": "1.3.1",
     "isomorphic-fetch": "3.0.0",
     "reflect-metadata": "0.1.13",
@@ -244,8 +244,8 @@
     "@nrwl/storybook": "14.5.10",
     "@nrwl/web": "14.5.10",
     "@nrwl/workspace": "14.5.10",
-    "@sentry/angular": "7.12.0",
-    "@sentry/tracing": "7.12.0",
+    "@sentry/angular": "7.12.1",
+    "@sentry/tracing": "7.12.1",
     "@storybook/addon-essentials": "6.5.10",
     "@storybook/addon-links": "6.5.10",
     "@storybook/addons": "6.5.10",
@@ -262,7 +262,7 @@
     "@types/google-protobuf": "3.15.6",
     "@types/graphql-upload": "8.0.11",
     "@types/hammerjs": "2.0.41",
-    "@types/jest": "28.1.8",
+    "@types/jest": "29.0.0",
     "@types/jsdom": "20.0.0",
     "@types/marked": "4.0.6",
     "@types/node": "18.7.14",
@@ -291,9 +291,9 @@
     "eslint-plugin-no-unsanitized": "4.0.1",
     "eslint-plugin-prettier": "4.2.1",
     "eslint-plugin-rxjs": "5.0.2",
-    "eslint-plugin-simple-import-sort": "7.0.0",
+    "eslint-plugin-simple-import-sort": "8.0.0",
     "eslint-plugin-unicorn": "43.0.2",
-    "firebase": "9.9.3",
+    "firebase": "9.9.4",
     "firebase-admin": "11.0.1",
     "firebase-functions": "3.23.0",
     "google-protobuf": "3.21.0",
@@ -315,7 +315,7 @@
     "ngx-markdown": "14.0.1",
     "nx": "14.5.10",
     "postcss": "8.4.16",
-    "postcss-import": "14.1.0",
+    "postcss-import": "15.0.0",
     "postcss-preset-env": "7.8.0",
     "postcss-url": "10.1.3",
     "prettier": "2.7.1",
