+++ tools/jest.config.ts
@@ -2,10 +2,10 @@ const config = {
   coverageDirectory: '../coverage/tools',
   coverageThreshold: {
     global: {
-      branches: 89,
-      functions: 95,
-      lines: 99,
-      statements: 99,
+      branches: 77,
+      functions: 85,
+      lines: 86,
+      statements: 84,
     },
   },
   displayName: 'tools',
